Flatten structure of file-layout inside the export dir.
Before this PR all data-types (files, versions, trashbin...)
were stored inside a single files directory which closely followed ownClouds
home-folder layout. This allowed for fast iteration but also coupled
the format to ownCloud which in turn introduced some design quirks
(/files/files/) in files.jsonl (#111)

All data type specific folders are now stored in the root directory of
the export which allows simpler mapping from metadata (#118).
This is also reflected in the architecture: The exporter traverses down from
the specific directories instead from the home.

A special "root folder" was introduced in files.jsonl to further decouple things
from owncloud and to be able to carry the e-tag for the whole tree. Instead of
"/files", "/" is now root in the export.

Path class has been added to reduce path-merging boilerplate.